+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add click-to-toggle behavior for expandable project tree and Running outline rows
+
 ## [0.1.2] - 2026-06-08
 
 ### Added